fix windows warnings

+++ intra_process_comms/src/two_node_pipeline/two_node_pipeline.cpp
@@ -25,7 +25,7 @@ struct Producer : public rclcpp::Node
   {
     pub_ = this->create_publisher<std_msgs::msg::Int32>(output, rmw_qos_profile_default);
     timer_ = this->create_wall_timer(1_s, [this]() {
-      static size_t count = 0;
+      static int32_t count = 0;
       std_msgs::msg::Int32::UniquePtr msg(new std_msgs::msg::Int32());
       msg->data = count++;
       printf("Published message with value: %d, and address: %p\n", msg->data, msg.get());
